+++ test/jasmine/tests/shapes_test.js
@@ -14,6 +14,7 @@ var customMatchers = require('../assets/custom_matchers');
 var createGraphDiv = require('../assets/create_graph_div');
 var destroyGraphDiv = require('../assets/destroy_graph_div');
 var failTest = require('../assets/fail_test');
+var drag = require('../assets/drag');
 
 
 describe('Test shapes defaults:', function() {
@@ -319,6 +320,23 @@ describe('Test shapes:', function() {
                 expect(countShapePathsInLowerLayer()).toEqual(0);
                 expect(countShapePathsInSubplots()).toEqual(0);
             })
+            .then(function() {
+                return Plotly.relayout(gd, {'shapes[0]': getRandomShape()});
+            })
+            .then(function() {
+                expect(countShapePathsInUpperLayer()).toEqual(1);
+                expect(countShapePathsInLowerLayer()).toEqual(0);
+                expect(countShapePathsInSubplots()).toEqual(0);
+                expect(gd.layout.shapes.length).toBe(1);
+
+                return Plotly.relayout(gd, {'shapes[0]': null});
+            })
+            .then(function() {
+                expect(countShapePathsInUpperLayer()).toEqual(0);
+                expect(countShapePathsInLowerLayer()).toEqual(0);
+                expect(countShapePathsInSubplots()).toEqual(0);
+                expect(gd.layout.shapes).toBeUndefined();
+            })
             .catch(failTest)
             .then(done);
         });
@@ -378,6 +396,71 @@ describe('Test shapes:', function() {
             .catch(failTest)
             .then(done);
         });
+    });
+});
+
+describe('shapes axis reference changes', function() {
+    'use strict';
+
+    var gd;
+
+    beforeEach(function(done) {
+        gd = createGraphDiv();
+
+        Plotly.plot(gd, [
+            {y: [1, 2, 3]},
+            {y: [1, 2, 3], yaxis: 'y2'}
+        ], {
+            yaxis: {domain: [0, 0.4]},
+            yaxis2: {domain: [0.6, 1]},
+            shapes: [{
+                xref: 'x', yref: 'paper', type: 'rect',
+                x0: 0.8, x1: 1.2, y0: 0, y1: 1,
+                fillcolor: '#eee', layer: 'below'
+            }]
+        }).then(done);
+    });
+
+    afterEach(destroyGraphDiv);
+
+    function getShape(index) {
+        var s = d3.selectAll('path[data-index="' + index + '"]');
+        expect(s.size()).toBe(1);
+        return s;
+    }
+
+    it('draws the right number of objects and updates clip-path correctly', function(done) {
+
+        expect(getShape(0).attr('clip-path') || '').toMatch(/x\)$/);
+
+        Plotly.relayout(gd, {
+            'shapes[0].xref': 'paper',
+            'shapes[0].x0': 0.2,
+            'shapes[0].x1': 0.6
+        })
+        .then(function() {
+            expect(getShape(0).attr('clip-path')).toBe(null);
+
+            return Plotly.relayout(gd, {
+                'shapes[0].yref': 'y2',
+                'shapes[0].y0': 1.8,
+                'shapes[0].y1': 2.2,
+            });
+        })
+        .then(function() {
+            expect(getShape(0).attr('clip-path') || '').toMatch(/^[^x]+y2\)$/);
+
+            return Plotly.relayout(gd, {
+                'shapes[0].xref': 'x',
+                'shapes[0].x0': 1.5,
+                'shapes[0].x1': 20
+            });
+        })
+        .then(function() {
+            expect(getShape(0).attr('clip-path') || '').toMatch(/xy2\)$/);
+        })
+        .catch(failTest)
+        .then(done);
     });
 });
 
@@ -748,7 +831,7 @@ describe('Test shapes', function() {
 
         var initialCoordinates = getShapeCoordinates(layoutShape, x2p, y2p);
 
-        return resize(direction, node, dx, dy).then(function() {
+        return drag(node, dx, dy, direction).then(function() {
             var finalCoordinates = getShapeCoordinates(layoutShape, x2p, y2p);
 
             var keyN, keyS, keyW, keyE;
@@ -815,118 +898,3 @@ describe('Test shapes', function() {
         return coordinates;
     }
 });
-
-var DBLCLICKDELAY = require('@src/plots/cartesian/constants').DBLCLICKDELAY;
-
-function mouseDown(node, x, y) {
-    node.dispatchEvent(new MouseEvent('mousedown', {
-        bubbles: true,
-        clientX: x,
-        clientY: y
-    }));
-}
-
-function mouseMove(node, x, y) {
-    node.dispatchEvent(new MouseEvent('mousemove', {
-        bubbles: true,
-        clientX: x,
-        clientY: y
-    }));
-}
-
-function mouseUp(node, x, y) {
-    node.dispatchEvent(new MouseEvent('mouseup', {
-        bubbles: true,
-        clientX: x,
-        clientY: y
-    }));
-}
-
-function drag(node, dx, dy) {
-    var bbox = node.getBoundingClientRect(),
-        fromX = (bbox.left + bbox.right) / 2,
-        fromY = (bbox.bottom + bbox.top) / 2,
-        toX = fromX + dx,
-        toY = fromY + dy;
-
-    mouseMove(node, fromX, fromY);
-    mouseDown(node, fromX, fromY);
-
-    var promise = waitForDragCover().then(function(dragCoverNode) {
-        mouseMove(dragCoverNode, toX, toY);
-        mouseUp(dragCoverNode, toX, toY);
-        return waitForDragCoverRemoval();
-    });
-
-    return promise;
-}
-
-function resize(direction, node, dx, dy) {
-    var bbox = node.getBoundingClientRect();
-
-    var fromX, fromY, toX, toY;
-
-    if(~direction.indexOf('n')) fromY = bbox.top;
-    else if(~direction.indexOf('s')) fromY = bbox.bottom;
-    else fromY = (bbox.bottom + bbox.top) / 2;
-
-    if(~direction.indexOf('w')) fromX = bbox.left;
-    else if(~direction.indexOf('e')) fromX = bbox.right;
-    else fromX = (bbox.left + bbox.right) / 2;
-
-    toX = fromX + dx;
-    toY = fromY + dy;
-
-    mouseMove(node, fromX, fromY);
-    mouseDown(node, fromX, fromY);
-
-    var promise = waitForDragCover().then(function(dragCoverNode) {
-        mouseMove(dragCoverNode, toX, toY);
-        mouseUp(dragCoverNode, toX, toY);
-        return waitForDragCoverRemoval();
-    });
-
-    return promise;
-}
-
-function waitForDragCover() {
-    return new Promise(function(resolve) {
-        var interval = DBLCLICKDELAY / 4,
-            timeout = 5000;
-
-        var id = setInterval(function() {
-            var dragCoverNode = d3.selectAll('.dragcover').node();
-            if(dragCoverNode) {
-                clearInterval(id);
-                resolve(dragCoverNode);
-            }
-
-            timeout -= interval;
-            if(timeout < 0) {
-                clearInterval(id);
-                throw new Error('waitForDragCover: timeout');
-            }
-        }, interval);
-    });
-}
-
-function waitForDragCoverRemoval() {
-    return new Promise(function(resolve) {
-        var interval = DBLCLICKDELAY / 4,
-            timeout = 5000;
-
-        var id = setInterval(function() {
-            var dragCoverNode = d3.selectAll('.dragcover').node();
-            if(!dragCoverNode) {
-                clearInterval(id);
-                resolve(dragCoverNode);
-            }
-
-            timeout -= interval;
-            if(timeout < 0) {
-                clearInterval(id);
-                throw new Error('waitForDragCoverRemoval: timeout');
-            }
-        }, interval);
-    });
-}